+++ src/LogLevel.js
@@ -1,0 +1,7 @@
+export default Object.freeze({
+  None: 'NONE',
+  Verbose: 'VERBOSE',
+  Debug: 'DEBUG',
+  Info: 'INFO',
+  Warn: 'WARN'
+});