+++ src/execution/FetchStrategy.js
@@ -1,7 +1,27 @@
-export default Object.freeze({
-  FetchFromCacheOrThrow: 'FETCH_FROM_CACHE_OR_THROW',
-  FetchFromCacheOrFallbackNetwork: 'FETCH_FROM_CACHE_OR_FALLBACK_NETWORK',
-  FetchFromCacheAndNetwork: 'FETCH_FROM_CACHE_AND_NETWORK',
-  FetchFromNetwork: 'FETCH_FROM_NETWORK',
-  FetchFromNetworkAndSkipCacheUpdates: 'FETCH_FROM_NETWORK_AND_SKIP_CACHE_UPDATES'
-});
+export default {
+  toString: function(fetchStrategy) {
+    switch(fetchStrategy) {
+      case 'FETCH_FROM_CACHE_OR_THROW':
+        return 'cache only';
+
+      case 'FETCH_FROM_CACHE_OR_FALLBACK_NETWORK':
+        return 'cache or network';
+
+      case 'FETCH_FROM_CACHE_AND_NETWORK':
+        return 'cache and network';
+
+      case 'FETCH_FROM_NETWORK':
+        return 'network only';
+
+      case 'FETCH_FROM_NETWORK_AND_NO_CACHE':
+        return 'network only and no caching';
+    }
+  },
+  ...Object.freeze({
+    FetchFromCacheOrThrow: 'FETCH_FROM_CACHE_OR_THROW',
+    FetchFromCacheOrFallbackNetwork: 'FETCH_FROM_CACHE_OR_FALLBACK_NETWORK',
+    FetchFromCacheAndNetwork: 'FETCH_FROM_CACHE_AND_NETWORK',
+    FetchFromNetwork: 'FETCH_FROM_NETWORK',
+    FetchFromNetworkAndNoCache: 'FETCH_FROM_NETWORK_AND_NO_CACHE'
+  })
+};