+++ src/execution/Notifier.js
@@ -1,4 +1,5 @@
-import { isEmptyObjectLiteral } from 'object-array-utils';
+import { isArray, isEmptyArray, isEmptyObjectLiteral } from 'object-array-utils';
+import Logger from '../Logger';
 
 export default class Notifier {
   static subscribers = new Set();
@@ -13,13 +14,25 @@ export default class Notifier {
     };
   }
 
-  static notify(data) {
-    if (isEmptyObjectLiteral(data)) {
+  static notify(entities) {
+    if (!isArray(entities)) {
+      throw new Error();
+    }
+
+    Logger.verbose(() => {
+      const entitiesWithoutMeta = [...entities].map((entity) => {
+        const { __meta, ...entityWithoutMeta } = entity;
+        return entityWithoutMeta;
+      });
+      return `Notifying ${Notifier.subscribers.size} subscribers with new entities: ${JSON.stringify(entitiesWithoutMeta, null, 2)}`;
+    });
+
+    if (isEmptyArray(entities)) {
       return;
     }
 
     for (const { subscriber } of Notifier.subscribers) {
-      subscriber.updateCache(data);
+      subscriber.updateCache(entities);
     }
   }
 }